Fail earlier when given a role that's neither an ARN nor an alias
Hopefully we save people some time :)

This change tests the incoming role name against a regex to determine whether
it's an ARN for an IAM role. If not, it's treated as an alias.

This removes the possibility of creating aliases that look like ARNs. I think
that's OK, since my imagination only gives me malicious uses cases for that.