cmd/root.go(rootCmd): Modify description to call it "command", not "subcommand".
To me, "subcommand" conjures something like the "s3" in "aws s3", that is, a subcommand of the tool
itself.  Whereas in fact, we mean to refer to external commands that yak will execute on our
behalf.  I think just calling it `<command>` in the docs and help text is clearer.