README.md: Update bit about '--'.
Technically, the '--' can be either before or after yak's role argument, in fact i'm in the habit of
running commands like `yak prod -- dothefoo` (which is a bit more GNU-opt-y i believe), whereas the
documentation implies this isn't right.  I've updated the copy to be a bit more general in that
regard.