cmd(root): Prefix errors to terminal with "yak: ".
Since `yak` is a shim, a user is almost always doing `yak foo command`, but if something goes wrong,
such as forgetting the `--` in `yak fruit-role banana --peeled`, it's pretty confusing to get an
error like "unknown flag --peeled", if the user knows for certain `banana` can be peeled.  This
small change prefixes yak's errors, so that the user knows at a glance what failed.