cli/login.go: Use pinentry, optionally.
Default to old password input, except with --pinentry mode.

We want to allow this option for when GUI apps call us in the background and we can't prompt for
password/TOTP via stdin/stdout as usual.  For now we're making it opt-in, so we don't "move too many
people's cheese".